+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bugfix
 
 - Change ImageGallery image scale from preview to large. @tisto
+- Also use `settings.internalApiPath` in url helpers `isInternalURL`, `flattenToAppUrl` and `flattenHTMLToAppURL` @tiberiuichim
 - Fix getBlocks helper when blocks_layout has no `items` (default PloneSite with no volto homepage)
 
 ### Internal